+++ app/gui/src/dashboard/components/AriaComponents/Text/Text.tsx
@@ -29,6 +29,7 @@ export interface TextProps
   readonly tooltipCrossOffset?: number
 }
 
+// eslint-disable-next-line react-refresh/only-export-components
 export const TEXT_STYLE = twv.tv({
   base: '',
   variants: {