+++ app/gui/src/dashboard/hooks/focusHooks.ts
@@ -6,10 +6,6 @@ import * as focusDirectionProvider from '#/providers/FocusDirectionProvider'
 
 import * as aria from '#/components/aria'
 
-// ==========================
-// === useHandleFocusMove ===
-// ==========================
-
 /**
  * The type of `react-aria` keyboard events. It must be extracted out of this type as it is not
  * exposed from the library itself.
@@ -62,10 +58,6 @@ export function useHandleFocusMove(direction: 'horizontal' | 'vertical') {
   )
 }
 
-// =========================
-// === useSoleFocusChild ===
-// =========================
-
 /**
  * Return JSX props to make a child focusable by `Navigator2D`. DOES NOT handle arrow keys,
  * because this hook assumes the child is the only focus child.
@@ -77,10 +69,6 @@ export function useSoleFocusChild() {
     className: focusChildClass,
   } satisfies React.HTMLAttributes<Element>
 }
-
-// =====================
-// === useFocusChild ===
-// =====================
 
 /**
  * Return JSX props to make a child focusable by `Navigator2D`, and make the child handle arrow